backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.9
pydantic-settings>=2.2.1
pytest>=8.1.1
httpx>=0.27.0
greenlet>=3.0.3
python-multipart>=0.0.9
email-validator>=2.1.1
cryptography>=42.0.5
passlib[bcrypt]>=1.7.4
